Add files via upload
Welcome to this repository! This repository contains the HTML and CSS code for Sticky NavBar based webpage. This project was developed using Visual Studio Code, a powerful and versatile source code editor.

Contents:

index.html: This file contains the main structure and content of the website.
styles.css: This file contains the CSS styling to enhance the visual presentation and layout of the website.